requirements.txt
aiogram==2.25.1
python-dotenv
pandas
fpdf2
qrcode
Pillow
requests
reportlab
